third_party/WebKit/LayoutTests/platform/mac-mac10.9/virtual/mojo-loading/http/tests/inspector/indexeddb/database-refresh-view-expected.txt

Tests refreshing the database information and data views.

Dumping IndexedDB tree:
    (empty)
Created database.
Dumping IndexedDB tree:
    database: Database1 - http://127.0.0.1:8000
        Object store: Store1
            (no indexes)
    database: testDatabase - http://127.0.0.1:8000
        (no object stores)
Created first objectstore.
Dumping IndexedDB tree:
    database: Database1 - http://127.0.0.1:8000
        Object store: Store1
            (no indexes)
    database: testDatabase - http://127.0.0.1:8000
        (no object stores)
Created second objectstore.
Dumping IndexedDB tree:
    database: Database1 - http://127.0.0.1:8000
        Object store: Store1
            (no indexes)
    database: testDatabase - http://127.0.0.1:8000
        (no object stores)
Error: Promise was collected
